requirements.txt
Flask
gunicorn
psycopg2-binary
phonenumbers
itsdangerous
requests
SQLAlchemy
Flask_SQLAlchemy
python-dotenv
flask-migrate
